requirements.txt
torch>=2.0
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
scikit-learn>=1.3
matplotlib>=3.7
